Add files via upload
This is my 4th python project which is simple but I consider this to be a fail. I didn't really know how I would make this so that it is playable but idrc I tried. I will continue making about 4 projects every day (if possible) so that I can improve.